requirements.txt
pybit==5.8.0
python-dotenv==1.1.0
pandas==2.2.3
loguru==0.7.3
pytelegrambotapi==4.27.0
numpy==2.2.0
ta==0.11.0